# openapi.yaml
openapi: 3.0.1
info:
    title: Pratica 4 API
    description: Pratica 4 API documentation
    termsOfService: ''
    contact:
        name: ''
        url: ''
        email: 'pedpao@pedpao.com'
    license:
        name: unlicensed
        url: ''
    version: 0.0.1
servers:
    - url: http://localhost:8080
      description: Generated server url
paths:
    /api/padarias/{id}:
        get:
            tags:
                - padaria-resource
            operationId: getPadaria
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/PadariaDTO'
        put:
            tags:
                - padaria-resource
            operationId: updatePadaria
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PadariaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/PadariaDTO'
        delete:
            tags:
                - padaria-resource
            operationId: deletePadaria
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                '200':
                    description: OK
        patch:
            tags:
                - padaria-resource
            operationId: partialUpdatePadaria
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PadariaDTO'
                    application/merge-patch+json:
                        schema:
                            $ref: '#/components/schemas/PadariaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/PadariaDTO'
...
    /api/padarias:
        get:
            tags:
                - padaria-resource
            operationId: getAllPadarias
            parameters:
                - name: page
                  in: query
                  description: Zero-based page index (0..N)
                  required: false
                  schema:
                      minimum: 0
                      type: integer
                      default: 0
                - name: size
                  in: query
                  description: The size of the page to be returned
                  required: false
                  schema:
                      minimum: 1
                      type: integer
                      default: 20
                - name: sort
                  in: query
                  description: >-
                      Sorting criteria in the format: property,(asc|desc). Default sort
                      order is ascending. Multiple sort criteria are supported.
                  required: false
                  schema:
                      type: array
                      items:
                          type: string
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/PadariaDTO'
        post:
            tags:
                - padaria-resource
            operationId: createPadaria
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PadariaDTO'
                required: true
            responses:
                '200':
                    description: OK
                    content:
                        '*/*':
                            schema:
                                $ref: '#/components/schemas/PadariaDTO'
...
components:
    schemas:
        PadariaDTO:
            required:
                - bairro
                - cep
                - cnpj
                - estado
                - fantasia
                - id
                - nome
                - numero
                - rua
                - telefone
            type: object
            properties:
                id:
                    minimum: 1
                    type: integer
                    format: int64
                nome:
                    maxLength: 255
                    minLength: 3
                    type: string
                fantasia:
                    maxLength: 255
                    minLength: 3
                    type: string
                telefone:
                    maxLength: 15
                    minLength: 10
                    type: string
                cnpj:
                    maxLength: 14
                    minLength: 14
                    type: string
                cep:
                    maxLength: 8
                    minLength: 8
                    type: string
                rua:
                    maxLength: 255
                    minLength: 3
                    type: string
                bairro:
                    maxLength: 255
                    minLength: 3
                    type: string
                estado:
                    maxLength: 2
                    minLength: 2
                    type: string
                numero:
                    maxLength: 255
                    minLength: 1
                    type: string
                complemento:
                    type: string
                contato:
                    type: string
                user:
                    $ref: '#/components/schemas/UserDTO'
        UserDTO:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                login:
                    type: string
...